code 리팩토링 중
    1. Controller는 웹 전달만 하는 계층
        비즈니스 로직은 서비스에서 처리할 수 있도록 하자.

    2. Service 에서 DTO <-> Entity 변환의 역할을 맡도록 하자.
        그렇지 않으면 LazyInitializationException 노출 위험이 커짐
        트랜잭션 안에서 하이버네이트의 명령이 이루어질 수 있도록 하자.

    3. 서비스 안에서 다른 서비스 호출 가능
        캘린더에서 그룹조회를 원한다면 당연히 그룹서비스를 불러 호출
        그룹 레파지토리 ㄴㄴ
        --> 이를 위해 dto 개념이 명확해야 할듯..

     4. RequestDTO / ResponseDto 구분
        요청과 응답 그리고 엔티티의 경계는 확실히 해두자.

+++ src/main/java/damdorani/dandulendar/dto/CalendarDetailForm.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotEmpty;
 @Getter
 @Setter
 public class CalendarDetailForm {
-    private int cal_dtl_id;
+    private Long cal_dtl_id;
     @NotEmpty(message = "일정명은 필수 입니다.")
     private String title;
     private String contents;
@@ -18,5 +18,5 @@ public class CalendarDetailForm {
     private String end_time;
     private String repeat_yn;
     private String allday_yn;
-    private int cal_id;
+    private Long cal_id;
 }
